requirements.txt
fastapi
uvicorn
websockets
twilio
certifi
aiomysql
python-dotenv
openai
cryptography
python-multipart
